int64 pks, not in32
* I am dumb. graphql will not ever be representing this ids as actual
int so do not limit them because of graphql spec. Make them int64 so
that running out is highly unlikely, but performance is better than
uuid.

+++ graphqlApi/root_resolvers.go
@@ -155,7 +155,7 @@ func (r *Resolver) Sensor(ctx context.Context, args struct{ ID graphql.ID }) (*s
 		log.Printf("%v for input: %v", err, spew.Sdump(args))
 		return nil, err
 	}
-	sensorId := int32(_sensorId) // does this need to happen or is passing an int64 in ok here?
+	sensorId := int64(_sensorId) // does this need to happen or is passing an int64 in ok here?
 
 	sensor, err := worrywort.FindSensor(map[string]interface{}{"id": sensorId, "user_id": *user.Id}, db)
 	if err != nil {
