+++ worrywort/sensor.go
@@ -12,10 +12,10 @@ import (
 // TODO: This may also want extra metadata such as model or type?  That is probably
 // going too far for now, so keep it simple.
 type Sensor struct {
-	Id        *int32 `db:"id"`
+	Id        *int64 `db:"id"`
 	Name      string `db:"name"`
 	CreatedBy *User  `db:"created_by,prefix=u"`
-	UserId    *int32 `db:"user_id"`
+	UserId    *int64 `db:"user_id"`
 
 	CreatedAt time.Time `db:"created_at"`
 	UpdatedAt time.Time `db:"updated_at"`
@@ -119,7 +119,7 @@ func SaveSensor(db *sqlx.DB, tm Sensor) (Sensor, error) {
 func InsertSensor(db *sqlx.DB, t Sensor) (Sensor, error) {
 	var updatedAt time.Time
 	var createdAt time.Time
-	sensorId := new(int32)
+	sensorId := new(int64)
 
 	query := db.Rebind(`INSERT INTO sensors (user_id, name, updated_at)
 		VALUES (?, ?, NOW()) RETURNING id, created_at, updated_at`)
